Fix plugin to pass absolute paths to CLI
The plugin was passing relative paths to the CLI, but the CLI might be running
from a different working directory. Now using vim.fn.fnamemodify with ':p' to
convert all project paths to absolute before passing to CLI.

Fixed in:
- Stage_UbtGenCmd (gen command)
- BuildCoroutine (build command)
- run command

This should fix the 'Unable to find project file' error when running from Neovim.